regulations created

+++ database/migrations/2024_07_07_100348_create_organizations_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('organizations', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->integer('stir');
+            $table->bigInteger('stir');
             $table->softDeletes();
             $table->timestamps();
         });
